refactor(doctor): extract safe-bin diagnostics helpers
